+++ packages/twenty-ui/src/navigation/index.ts
@@ -18,41 +18,41 @@ export { UndecoratedLink } from './link/components/UndecoratedLink';
 export { CAL_LINK } from './link/constants/Cal';
 export { GITHUB_LINK } from './link/constants/GithubLink';
 export { TWENTY_PRICING_LINK } from './link/constants/TwentyPricingLink';
+export type { MenuPickerProps } from './menu/components/MenuPicker';
+export { MenuPicker } from './menu/components/MenuPicker';
 export type {
   MenuItemIconButton,
   MenuItemProps,
-} from './menu-item/components/MenuItem';
-export { MenuItem } from './menu-item/components/MenuItem';
-export type { MenuItemAvatarProps } from './menu-item/components/MenuItemAvatar';
-export { MenuItemAvatar } from './menu-item/components/MenuItemAvatar';
-export type { MenuItemCommandProps } from './menu-item/components/MenuItemCommand';
-export { MenuItemCommand } from './menu-item/components/MenuItemCommand';
-export type { MenuItemCommandHotKeysProps } from './menu-item/components/MenuItemCommandHotKeys';
-export { MenuItemCommandHotKeys } from './menu-item/components/MenuItemCommandHotKeys';
-export type { MenuItemDraggableProps } from './menu-item/components/MenuItemDraggable';
-export { MenuItemDraggable } from './menu-item/components/MenuItemDraggable';
-export { MenuItemMultiSelect } from './menu-item/components/MenuItemMultiSelect';
-export { MenuItemMultiSelectAvatar } from './menu-item/components/MenuItemMultiSelectAvatar';
-export { MenuItemMultiSelectTag } from './menu-item/components/MenuItemMultiSelectTag';
-export type { MenuItemNavigateProps } from './menu-item/components/MenuItemNavigate';
-export { MenuItemNavigate } from './menu-item/components/MenuItemNavigate';
-export type { MenuItemPickerProps } from './menu-item/components/MenuItemPicker';
-export { MenuItemPicker } from './menu-item/components/MenuItemPicker';
+} from './menu/menu-item/components/MenuItem';
+export { MenuItem } from './menu/menu-item/components/MenuItem';
+export type { MenuItemAvatarProps } from './menu/menu-item/components/MenuItemAvatar';
+export { MenuItemAvatar } from './menu/menu-item/components/MenuItemAvatar';
+export type { MenuItemCommandProps } from './menu/menu-item/components/MenuItemCommand';
+export { MenuItemCommand } from './menu/menu-item/components/MenuItemCommand';
+export type { MenuItemCommandHotKeysProps } from './menu/menu-item/components/MenuItemCommandHotKeys';
+export { MenuItemCommandHotKeys } from './menu/menu-item/components/MenuItemCommandHotKeys';
+export type { MenuItemDraggableProps } from './menu/menu-item/components/MenuItemDraggable';
+export { MenuItemDraggable } from './menu/menu-item/components/MenuItemDraggable';
+export { MenuItemMultiSelect } from './menu/menu-item/components/MenuItemMultiSelect';
+export { MenuItemMultiSelectAvatar } from './menu/menu-item/components/MenuItemMultiSelectAvatar';
+export { MenuItemMultiSelectTag } from './menu/menu-item/components/MenuItemMultiSelectTag';
+export type { MenuItemNavigateProps } from './menu/menu-item/components/MenuItemNavigate';
+export { MenuItemNavigate } from './menu/menu-item/components/MenuItemNavigate';
 export {
   StyledMenuItemSelect,
   MenuItemSelect,
-} from './menu-item/components/MenuItemSelect';
-export { MenuItemSelectAvatar } from './menu-item/components/MenuItemSelectAvatar';
+} from './menu/menu-item/components/MenuItemSelect';
+export { MenuItemSelectAvatar } from './menu/menu-item/components/MenuItemSelectAvatar';
 export {
   colorLabels,
   MenuItemSelectColor,
-} from './menu-item/components/MenuItemSelectColor';
-export { MenuItemSelectTag } from './menu-item/components/MenuItemSelectTag';
-export type { MenuItemSuggestionProps } from './menu-item/components/MenuItemSuggestion';
-export { MenuItemSuggestion } from './menu-item/components/MenuItemSuggestion';
-export { MenuItemToggle } from './menu-item/components/MenuItemToggle';
-export { MenuItemLeftContent } from './menu-item/internals/components/MenuItemLeftContent';
-export type { MenuItemBaseProps } from './menu-item/internals/components/StyledMenuItemBase';
+} from './menu/menu-item/components/MenuItemSelectColor';
+export { MenuItemSelectTag } from './menu/menu-item/components/MenuItemSelectTag';
+export type { MenuItemSuggestionProps } from './menu/menu-item/components/MenuItemSuggestion';
+export { MenuItemSuggestion } from './menu/menu-item/components/MenuItemSuggestion';
+export { MenuItemToggle } from './menu/menu-item/components/MenuItemToggle';
+export { MenuItemLeftContent } from './menu/menu-item/internals/components/MenuItemLeftContent';
+export type { MenuItemBaseProps } from './menu/menu-item/internals/components/StyledMenuItemBase';
 export {
   StyledMenuItemBase,
   StyledMenuItemLabel,
@@ -62,8 +62,8 @@ export {
   StyledDraggableItem,
   StyledHoverableMenuItemBase,
   StyledMenuItemIconCheck,
-} from './menu-item/internals/components/StyledMenuItemBase';
-export type { MenuItemAccent } from './menu-item/types/MenuItemAccent';
+} from './menu/menu-item/internals/components/StyledMenuItemBase';
+export type { MenuItemAccent } from './menu/menu-item/types/MenuItemAccent';
 export { NavigationBar } from './navigation-bar/components/NavigationBar';
 export { NavigationBarItem } from './navigation-bar/components/NavigationBarItem';
 export { NotificationCounter } from './notification-counter/components/NotificationCounter';
