+++ packages/twenty-ui/src/navigation/menu/components/index.ts
@@ -1,0 +1,1 @@
+export * from './MenuPicker';
