+++ packages/twenty-ui/src/navigation/menu/index.ts
@@ -1,0 +1,2 @@
+export * from './components';
+export * from './menu-item';
