+++ packages/twenty-ui/src/navigation/menu/menu-item/components/MenuItem.tsx
