+++ packages/twenty-ui/src/navigation/menu/menu-item/components/MenuItemSelectTag.tsx
