+++ packages/twenty-ui/src/navigation/menu/menu-item/internals/components/StyledMenuItemBase.tsx
