+++ packages/twenty-ui/src/navigation/menu/menu-item/types/MenuItemAccent.ts
